refactor(cli): extract parser global args
